Add Skill relation to People

+++ Module/Controller/Adminhtml/Skill/Delete.php
@@ -10,8 +10,9 @@ use Magento\Backend\App\Action\Context;
 use Magento\Backend\Model\View\Result\Redirect;
 use Magento\Framework\App\Action\HttpGetActionInterface;
 use Magento\Framework\Controller\ResultFactory;
-use Vendor\Module\Model\SkillFactory;
+use Vendor\Module\Model\ResourceModel\PeopleSkill;
 use Vendor\Module\Model\ResourceModel\Skill as SkillResource;
+use Vendor\Module\Model\SkillFactory;
 
 class Delete extends Action implements HttpGetActionInterface
 {
@@ -23,13 +24,14 @@ class Delete extends Action implements HttpGetActionInterface
      * @param Context $context
      * @param SkillFactory $skillFactory
      * @param SkillResource $skillResource
+     * @param PeopleSkill $peopleSkillResource
      */
     public function __construct(
-        private readonly Context          $context,
+        private readonly Context       $context,
         private readonly SkillFactory  $skillFactory,
-        private readonly SkillResource $skillResource
-    )
-    {
+        private readonly SkillResource $skillResource,
+        private readonly PeopleSkill   $peopleSkillResource
+    ) {
         parent::__construct($context);
     }
 
@@ -45,6 +47,17 @@ class Delete extends Action implements HttpGetActionInterface
             $skill = $this->skillFactory->create();
             $this->skillResource->load($skill, $skillId);
             if ($skill->getData('skill_id')) {
+
+                $relatedPeopleIds = $this->peopleSkillResource->getPeopleIds((int)$skillId);
+                if (!empty($relatedPeopleIds)) {
+                    $this->messageManager->addErrorMessage(
+                        __('This skill cannot be deleted because it is associated with one or more people.')
+                    );
+                    /** @var Redirect $redirect */
+                    $redirect = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
+                    return $redirect->setPath('*/*');
+                }
+
                 $this->skillResource->delete($skill);
                 $this->messageManager->addSuccessMessage(__('The record has been deleted.'));
             } else {
